web/bypassme/requirements.txt
click==8.1.3
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
Pygments==2.14.0
python-dotenv==1.0.0
rich==13.3.2
SQLAlchemy==2.0.7
typing_extensions==4.5.0
Werkzeug==2.2.3